ANDROID: staging: ion: Allow for attach and detach ops to be overridden
The ION core is supposed to allow for all of the dma-buf ops
to be overridden by a heap's dma-buf ops. However, the
attach and detach ops are not overridable, which doesn't make
much sense, as the other dma-buf ops assume that the attachment
structure that they operate on was allocated by their attach
callbacks, which aren't being invoked. Thus, allow for the dma-buf
attach and detach ops to be overridable.

Bug: 152270037
Change-Id: I981973911b8649700901740456e03ed4a1731c9a
Signed-off-by: Isaac J. Manjarres <isaacm@codeaurora.org>
Signed-off-by: Matthias Maennich <maennich@google.com>